pyproject.toml
[project]
name = "fpfh-icp-benchmarks"
version = "0.1.0"
description = "Benchmarking classic 6D pose estimation algorithms on BOP datasets"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
keywords = ["6d-pose-estimation", "computer-vision", "point-cloud", "fpfh", "icp", "bop"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "open3d>=0.17.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pyyaml>=6.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.62.0",
    "imageio>=2.35.1",
    "pypng>=0.20220715.0",
    "scikit-image>=0.21.0",
    "vispy>=0.14.2",
    "pyopengl>=3.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "isort>=5.0",
    "flake8>=3.8",
    "mypy>=0.910",
]

[project.scripts]
fpfh-benchmark = "scripts.run_benchmark:main"

[project.urls]
Homepage = "https://github.com/fedor-chervinskii/fpfh-icp-benchmarks"
Repository = "https://github.com/fedor-chervinskii/fpfh-icp-benchmarks"
Issues = "https://github.com/fedor-chervinskii/fpfh-icp-benchmarks/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | bop_toolkit
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
]
